reddit_wordcloud/requirements.txt
requests
boto3
matplotlib
wordcloud
Pillow